pyproject.toml
[tool.poetry]
name = "django-gql"
version = "0.0.1"
description = ""
authors = ["syfun <sunyu418@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/syfun/django-graphql"

packages = [
    { include = "djgql" }
]

[tool.poetry.dependencies]
python = ">=3.7"
python-gql = ">=0.1"
django = "^3"


[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
isort = "^4.3.21"
pre-commit = "^2.2.0"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
